Restore data tracing (right click on orange fields)

+++ SFEngine/SFCFF/CTG/Category2028.cs
@@ -12,7 +12,7 @@ namespace SFEngine.SFCFF.CTG
     {
         public ushort ArmyUnitID;
         public byte ResourceType;
-        public ushort ResourceValue;
+        public byte ResourceValue;
 
         public int GetID() => ArmyUnitID;
         public void SetID(int id) => ArmyUnitID = (ushort)id;
